requirements.txt

flask
python-telegram-bot==20.3
